add userComponent
